@@ -6,7 +6,7 @@
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#  http://www.apache.org/licenses/LICENSE-2.0
+#   http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -15,52 +15,78 @@
 # limitations under the License.
 #
 
+#
+# This needs to be in sync with the boot.rb. First we activate 
+# specific versions for the gems, and afterwards we require all gems. 
+# That's the best workaround for avoiding the usage of bundler. There 
+# is the rake task `crowbar:generate:dependencies` that updates the
+# dependencies within the boot file.
+#
+
 source "https://rubygems.org"
+gem "rails", "4.1.0"
 
-gem "rails", "2.3.17"
-gem "rack", "1.1.6"
-
-gem "app_config", "1.0.2"
-gem "chef", "10.24.4"
-gem "haml", "3.1.6"
-gem "hike", "1.2.1"
-gem "i18n", "0.4.2"
-gem "json", "1.6.1"
-gem "kwalify", "0.7.2"
-gem "multi_json", "1.0.3"
-gem "sass", "3.2.12"
-gem "simple-navigation", "3.7.0"
-gem "sprockets", "2.10.1"
-gem "sprockets-sass", "1.0.2"
-gem "sprockets-helpers", "1.1.0"
-gem "sqlite3", "1.3.6"
-gem "syslogger", "1.3.0"
-gem "tilt", "1.3.3"
-gem "mime-types", "1.18", :require => "mime/types"
+gem "dotenv-rails", "~> 0.10.0"
+gem "haml-rails", "~> 0.5.3"
+gem "sass-rails", "~> 4.0.3"
+gem "rack", "~> 1.5.2"
+gem "bcrypt", "~> 3.1.7"
+gem "kwalify", "~> 0.7.2"
+gem "simple-navigation", "~> 3.12.2"
+gem "simple_navigation_renderers", "~> 1.0.2"
+gem "sqlite3", "~> 1.3.9"
+gem "gli", "~> 2.9.0"
+gem "cocaine", "~> 0.5.4"
+gem "hashie", "~> 2.0.5"
+gem "delayed_job_active_record", "~> 4.0.1"
+gem "rainbows-rails", "~> 1.0.1"
+gem "redcarpet", "~> 3.1.1"
+gem "nokogiri", "~> 1.6.1"
+gem "bootstrap-sass", "~> 3.1.1"
+gem "font-awesome-rails", "~> 4.0.3"
+gem "closure-compiler", "~> 1.1.10"
+gem "cells", "~> 3.10.0"
+gem "js-routes", "~> 0.9.7"
+gem "json", "~> 1.8.1"
+gem "pry-rails", "~> 0.3.2"
+gem "activeresource", "~> 4.0.0", require: "active_resource"
+gem "activerecord-session_store", "~> 0.1.0", require: "activerecord/session_store"
 
 # We are not adding any version here because the SLES and Ubuntu chef 
 # packaging differ for the version, but we need this to load the chef 
 # gem correctly within the initializer.
-gem "chef"
+gem "chef", "~> 10.32.2"
+
+group :development, :test do
+  platforms :ruby_20 do
+    gem "byebug", "~> 2.7.0"
+  end
+
+  gem "rspec-rails", "~> 2.14.1"
+end
 
 group :development do
-  gem "brakeman", "1.8.2"
-  gem "rspec-rails", "1.3.4"
+  platforms :ruby do
+    gem "binding_of_caller", "~> 0.7.2"
+  end
+
+  gem "gemnasium-parser", "~> 0.1.9"
+  gem "guard", "~> 2.6.0"
+  gem "better_errors", "~> 1.1.0"
+  gem "meta_request", "~> 0.2.8"
+  gem "quiet_assets", "~> 1.0.2"
+  gem "brakeman", "~> 2.4.3", require: false
+  gem "rubocop", "~> 0.18.1", require: false
 end
 
 group :test do
-  gem "rspec-rails", "1.3.4"
-  gem "mocha", "0.10.5"
-  gem "sinatra", "1.2.9"
-  gem "webmock", "1.16.1"
-  gem "rcov", "1.0.0"
+  gem "mocha", "~> 1.0.0"
+  gem "webmock", "~> 1.17.4"
+  gem "sinatra", "~> 1.4.4"
+  gem "cucumber-rails", "~> 1.4.0", require: false
+  gem "simplecov", "~> 0.8.2", require: false
 end
 
-# Install gems from each barclamp. The crowbar_path variable is
-# needed by Gemfile.d/*.gemfile so that they can determine the correct
-# path at run-time rather than having to hard-code a path at
-# install-time which may be different and therefore incorrect.
-crowbar_path = File.expand_path("..", __FILE__)
-Dir.glob(File.join(crowbar_path, "Gemfile.d", "*.gemfile")) do |gemfile|
-  eval(IO.read(gemfile), binding)
+group :doc do
+  gem "sdoc", "~> 0.4.0", require: false
 end
