@@ -1,11 +1,13 @@
+# -*- encoding : utf-8 -*-
+#
 # Copyright 2011-2013, Dell
-# Copyright 2013, SUSE LINUX Products GmbH
+# Copyright 2013-2014, SUSE LINUX Products GmbH
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#  http://www.apache.org/licenses/LICENSE-2.0
+#   http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,39 +15,40 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-# Author: Rob Hirschfeld
-# Author: SUSE LINUX Products GmbH
-#
 
 module ApplicationHelper
-  include Sprockets::Helpers
-
   # Check if we are using a quite old bad internet explorer, currently used for
   # disableing drag and drop for this old browser
   def bad_explorer?
-    request.env['HTTP_USER_AGENT'].downcase =~ /msie ([1-8])/
+    @bad_explorer ||= request.env['HTTP_USER_AGENT'].downcase =~ /msie ([1-8])/
   end
 
   # Check if we are running on a suse system
   def suse_system?
-    @@suse_system ||= File.exist?("/etc/SuSE-release")
-  end
-
-  # Added this helper method to access app config, maybe we need some wrapping
-  # and it feels better to call a method instead of a class
-  def app_config
-    AppConfig
+    @suse_system ||= File.exist?("/etc/SuSE-release")
   end
 
   # A simple wrapper to access the branding configuration directly, looks much
   # cleaner within the views
   def branding_config
-    app_config[:branding]
+    @branding_config ||= begin
+      config = YAML::load_file(
+        Rails.root.join("config", "branding.yml")
+      )
+
+      Hashie::Mash.new config
+    rescue
+      Hashie::Mash.new
+    end
   end
 
   # Generate the meta title that gets displayed on the page meta information
   def meta_title
-    title = [branding_config[:page_title], branding_config[:page_slogan]].compact.join(" ")
+    title = [
+      branding_config.page_title, 
+      branding_config.page_slogan
+    ].compact.join(" ").strip
+
     "#{title}: #{controller.action_name.titleize}"
   end
 
@@ -53,9 +56,34 @@ module ApplicationHelper
   # stylesheets to get the stylesheets more dynamic included
   def registered_stylesheets
     @registered_stylesheets ||= begin
-      [
+      files = [
         "application"
       ]
+
+      barclamp_name = if @service_object
+        @service_object.barclamp rescue "crowbar"
+      else
+        "crowbar"
+      end
+
+      if barclamp_name == "barclamp"
+        barclamp_name = "crowbar"
+      end
+
+      barclamp_path = Rails.root.join(
+        "app",
+        "assets",
+        "stylesheets",
+        "barclamps",
+        barclamp_name,
+        "application.css.scss"
+      )
+
+      if barclamp_path.file?
+        files.push "barclamps/#{barclamp_name}/application"
+      end
+
+      files
     end
   end
 
@@ -69,9 +97,34 @@ module ApplicationHelper
   # javascripts to get the javascripts more dynamic included
   def registered_javascripts
     @registered_javascripts ||= begin
-      [
+      files = [
         "application"
       ]
+
+      barclamp_name = if @service_object
+        @service_object.barclamp rescue "crowbar"
+      else
+        "crowbar"
+      end
+
+      if barclamp_name == "barclamp"
+        barclamp_name = "crowbar"
+      end
+
+      barclamp_path = Rails.root.join(
+        "app",
+        "assets",
+        "javascripts",
+        "barclamps",
+        barclamp_name,
+        "application.js"
+      )
+
+      if barclamp_path.file?
+        files.push "barclamps/#{barclamp_name}/application"
+      end
+
+      files
     end
   end
 
@@ -86,18 +139,18 @@ module ApplicationHelper
     [].tap do |output|
       output.push content_tag(
         :span,
-        branding_config[:page_title],
+        branding_config.page_title,
         :class => "title"
       )
 
-      unless branding_config[:page_slogan].empty?
+      unless branding_config.page_slogan.empty?
         output.push content_tag(
           :span,
-          branding_config[:page_slogan],
+          branding_config.page_slogan,
           :class => "slogan"
         )
       end
-    end.join("\n")
+    end.join("\n").html_safe
   end
 
   # Include required meta tags like csrf token, viewport and such stuff
@@ -119,21 +172,7 @@ module ApplicationHelper
         :name => "viewport",
         :content => "width=device-width, initial-scale=1.0"
       )
-
-      if protect_against_forgery?
-        output.push tag(
-          :meta,
-          :name => "csrf-param",
-          :content => Rack::Utils.escape_html(request_forgery_protection_token)
-        )
-
-        output.push tag(
-          :meta,
-          :name => "csrf-token",
-          :content => Rack::Utils.escape_html(form_authenticity_token)
-        )
-      end
-    end.join("\n")
+    end.join("\n").html_safe
   end
 
   def have_openstack
@@ -159,7 +198,7 @@ module ApplicationHelper
   end
 
   def flash_for(value)
-    case value
+    case value.to_sym
     when :notice
       "success"
     when :alert
@@ -173,11 +212,11 @@ module ApplicationHelper
     if value.nil? or value.empty?
       content_tag(
         :span,
-        "&mdash;",
-        :class => "empty"
+        "&mdash;".html_safe,
+        class: "empty"
       )
     else
-      value
+      value.html_safe
     end
   end
 
@@ -185,17 +224,17 @@ module ApplicationHelper
     is_empty = if condition.nil?
       value.nil? or value.empty?
     else
-      ! condition
+      not condition
     end
 
     if is_empty
       content_tag(
         :span,
         fallback,
-        :class => "empty"
-      )
+        class: "empty"
+      ).html_safe
     else
-      value
+      value.html_safe
     end
   end
 
