@@ -1,4 +1,7 @@
-# Copyright 2011, Dell
+#!/usr/bin/env ruby
+#
+# Copyright 2011-2013, Dell
+# Copyright 2013-2014, SUSE LINUX Products GmbH
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -11,7 +14,13 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-# 
+#
+
+#
+#
+# TODO: Needs to be replaced by https://github.com/basecamp/sub
+#
+#
 
 require 'rubygems'
 require 'net/http'
@@ -67,20 +76,21 @@ end
   "delete" => [ "proposal_delete ARGV.shift", "delete <name> - delete a proposal" ],
   "commit" => [ "proposal_commit ARGV.shift", "commit <name> - Commit a proposal to active" ],
   "dequeue" => [ "proposal_dequeue ARGV.shift", "dequeue <name> - Dequeue a proposal to active" ]
+  #
+  # TBD
+  # * status
+  # * deactivate
+  #
 }
 
 @commands = {
   "help" => [ "help", "help - this page" ],
   "api_help" => [ "api_help", "crowbar API help - help for this barclamp." ],
-  "list" => [ "list", "list - show a list of current configs" ],
-  "show" => [ "show ARGV.shift", "show <name> - show a specific config" ],
-  "delete" => [ "delete ARGV.shift", "delete <name> - delete a config" ],
   "proposal" => [ "run_sub_command(@proposal_commands, ARGV.shift)", "proposal - Proposal sub-commands", @proposal_commands ],
   "elements" => [ "elements", "elements - List elements of a #{@barclamp} deploy" ],
-  "element_node" => [ "element_node ARGV.shift", "element_node <name> - List nodes that could be that element" ],
+  "element" => [ "element_node ARGV.shift", "element_node <name> - List nodes that could be that element" ],
   "transition" => [ "transition(ARGV.shift,ARGV.shift)", "transition <name> <state> - Transition machine named name to state" ]
 }
-
 
 def print_commands(cmds, spacer = "  ")
   cmds.each do |key, command|
@@ -137,14 +147,25 @@ def get_json(path)
   uri = URI.parse("http://#{@hostname}:#{@port}/crowbar/#{@barclamp}/1.0#{path}")
   res = authenticate(Net::HTTP::Get,uri)
 
-  puts "DEBUG: (g) hostname: #{uri.host}:#{uri.port}" if @debug
-  puts "DEBUG: (g) request: #{uri.path}" if @debug
-  puts "DEBUG: (g) return code: #{res.code}" if @debug
-  puts "DEBUG: (g) return body: #{res.body}" if @debug
+  puts "DEBUG: (get) hostname: #{uri.host}:#{uri.port}" if @debug
+  puts "DEBUG: (get) request: #{uri.path}" if @debug
+  puts "DEBUG: (get) return code: #{res.code}" if @debug
+  puts "DEBUG: (get) return body: #{res.body}" if @debug
 
   return [res.body, res.code.to_i ] if res.code.to_i != 200
 
-  struct = JSON.parse(res.body)
+  struct = case res
+  when Net::HTTPOK
+    {}
+  else
+    JSON.parse(res.body)
+  end
+
+  struct = begin
+    JSON.parse(res.body)
+  rescue
+    nil
+  end
 
   puts "DEBUG: (g) JSON parse structure = #{struct.inspect}" if @debug
 
@@ -157,7 +178,7 @@ def post_json(path, data)
 
   puts "DEBUG: (post) hostname: #{uri.host}:#{uri.port}" if @debug
   puts "DEBUG: (post) request: #{uri.path}" if @debug
-  puts "DEBUG: (post) data: #{@data}" if @debug
+  puts "DEBUG: (post) data: #{data}" if @debug
   puts "DEBUG: (post) return code: #{res.code}" if @debug
   puts "DEBUG: (post) return body: #{res.body}" if @debug
 
@@ -170,7 +191,7 @@ def put_json(path, data)
 
   puts "DEBUG: (put) hostname: #{uri.host}:#{uri.port}" if @debug
   puts "DEBUG: (put) request: #{uri.path}" if @debug
-  puts "DEBUG: (put) data: #{@data}" if @debug
+  puts "DEBUG: (put) data: #{data}" if @debug
   puts "DEBUG: (put) return code: #{res.code}" if @debug
   puts "DEBUG: (put) return body: #{res.body}" if @debug
 
@@ -179,32 +200,14 @@ end
 
 def delete_json(path)
   uri = URI.parse("http://#{@hostname}:#{@port}/crowbar/#{@barclamp}/1.0#{path}")
-  res = authenticate(Net::HTTP::Delete,uri)
+  res = authenticate(Net::HTTP::Get,uri)
 
-  puts "DEBUG: (d) hostname: #{uri.host}:#{uri.port}" if @debug
-  puts "DEBUG: (d) request: #{uri.path}" if @debug
-  puts "DEBUG: (d) return code: #{res.code}" if @debug
-  puts "DEBUG: (d) return body: #{res.body}" if @debug
+  puts "DEBUG: (delete) hostname: #{uri.host}:#{uri.port}" if @debug
+  puts "DEBUG: (delete) request: #{uri.path}" if @debug
+  puts "DEBUG: (delete) return code: #{res.code}" if @debug
+  puts "DEBUG: (delete) return body: #{res.body}" if @debug
 
   [res.body, res.code.to_i ]
-end
-
-
-def list
-  struct = get_json("/")
-
-  if struct[1] != 200
-    [ "Failed to talk to service list: #{struct[1]}: #{struct[0]}", 1 ]
-  elsif struct[0].nil? or struct[0].empty? 
-    [ "No current configurations", 0 ]
-  else
-    out = ""
-    struct[0].each do |name|
-      out = out + "\n" if out != ""
-      out = out + "#{name}"
-    end
-    [ out, 0 ]
-  end
 end
 
 def api_help
@@ -218,36 +221,8 @@ def api_help
   end
 end
 
-def show(name)
-  usage -1 if name.nil? or name == ""
-
-  struct = get_json("/#{name}")
-
-  if struct[1] == 200
-    [ "#{JSON.pretty_generate(struct[0])}", 0 ]
-  elsif struct[1] == 404
-    [ "No current configuration for #{name}", 1 ]
-  else
-    [ "Failed to talk to service show: #{struct[1]}: #{struct[0]}", 1 ]
-  end
-end
-
-def delete(name)
-  usage -1 if name.nil? or name == ""
- 
-  struct = delete_json("/#{name}")
-
-  if struct[1] == 200
-    [ "Deleted #{name}", 0 ]
-  elsif struct[1] == 404
-    [ "Delete failed for #{name}: Not Found", 1 ] 
-  else
-    [ "Failed to talk to service delete: #{struct[1]}: #{struct[0]}", 1 ]
-  end
-end
-
 def proposal_list
-  struct = get_json("/proposals/")
+  struct = get_json("/proposals.json")
 
   if struct[1] != 200
     [ "Failed to talk to service proposal list: #{struct[1]}: #{struct[0]}", 1 ]
@@ -266,7 +241,7 @@ end
 def proposal_show(name)
   usage -1 if name.nil? or name == ""
 
-  struct = get_json("/proposals/#{name}")
+  struct = get_json("/proposals/#{name}.json")
 
   if struct[1] == 200
     [ "#{JSON.pretty_generate(struct[0])}", 0 ]
@@ -282,7 +257,7 @@ def proposal_create(name)
 
   @data = "{\"id\":\"#{name}\"}" if @data.nil? or @data == ""
 
-  struct = put_json("/proposals", @data)
+  struct = put_json("/proposals.json", @data)
 
   if struct[1] == 200
     [ "Created #{name}", 0 ]
@@ -295,7 +270,7 @@ def proposal_edit(name)
   usage -1 if name.nil? or name == ""
 
   if @data.nil? or @data == ""
-    struct = get_json("/proposals/#{name}")
+    struct = get_json("/proposals/#{name}.json")
 
     if struct[1] == 200
       require 'tempfile'
@@ -312,7 +287,6 @@ def proposal_edit(name)
 
       begin
         file.open
-        #@data = JSON.pretty_generate(file.read)
         @data = JSON.pretty_generate(JSON.parse(file.read))
       ensure
         file.close
@@ -325,7 +299,7 @@ def proposal_edit(name)
     end
   end
 
-  struct = post_json("/proposals/#{name}", @data)
+  struct = post_json("/proposals/#{name}.json", @data)
 
   if struct[1] == 200
     [ "Edited #{name}", 0 ]
@@ -341,7 +315,7 @@ end
 def proposal_delete(name)
   usage -1 if name.nil? or name == ""
 
-  struct = delete_json("/proposals/#{name}")
+  struct = delete_json("/proposals/delete/#{name}.json")
 
   if struct[1] == 200
     [ "Deleted #{name}", 0 ]
@@ -355,7 +329,7 @@ end
 def proposal_commit(name)
   usage -1 if name.nil? or name == ""
 
-  struct = post_json("/proposals/commit/#{name}", @data)
+  struct = get_json("/proposals/commit/#{name}.json")
 
   if struct[1] == 200
     [ "Committed #{name}", 0 ]
@@ -369,7 +343,7 @@ end
 def proposal_dequeue(name)
   usage -1 if name.nil? or name == ""
 
-  struct = delete_json("/proposals/dequeue/#{name}")
+  struct = delete_json("/proposals/dequeue/#{name}.json")
 
   if struct[1] == 200
     [ "Dequeued #{name}", 0 ]
@@ -379,7 +353,7 @@ def proposal_dequeue(name)
 end
 
 def elements
-  struct = get_json("/elements")
+  struct = get_json("/elements.json")
 
   if struct[1] != 200
     [ "Failed to talk to service elements: #{struct[1]}: #{struct[0]}", 1 ]
@@ -398,7 +372,7 @@ end
 def element_node(element)
   usage -1 if element.nil? or element == ""
 
-  struct = get_json("/elements/#{element}")
+  struct = get_json("/elements/#{element}.json")
 
   if struct[1] != 200
     [ "Failed to talk to service element_node: #{struct[1]}: #{struct[0]}", 1 ]
@@ -421,7 +395,7 @@ def transition(name, state)
     "name" => name,
     "state" => state
   }
-  struct = post_json("/transition/default", data.to_json)
+  struct = post_json("/transition/default.json", data.to_json)
 
   if struct[1] == 200
     [ "Transitioned #{name}", 0 ]
