@@ -1,5 +1,7 @@
-#!/usr/bin/ruby
-# Copyright 2011, Dell
+#!/usr/bin/env ruby
+#
+# Copyright 2011-2013, Dell
+# Copyright 2013-2014, SUSE LINUX Products GmbH
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,6 +14,12 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+#
+
+#
+#
+# TODO: Needs to be replaced by https://github.com/basecamp/sub
+#
 #
 
 require 'rubygems'
