@@ -1,11 +1,13 @@
+# -*- encoding : utf-8 -*-
+#
 # Copyright 2011-2013, Dell
-# Copyright 2013, SUSE LINUX Products GmbH
+# Copyright 2013-2014, SUSE LINUX Products GmbH
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#  http://www.apache.org/licenses/LICENSE-2.0
+#   http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,13 +15,20 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-# Author: Dell Crowbar Team
-# Author: SUSE LINUX Products GmbH
-#
 
 class CrowbarController < BarclampController
-  def initialize
-    @service_object = CrowbarService.new(logger)
+  def index
+    @title = I18n.t("title", scope: "barclamp.crowbar.index")
+    super
+  end
+
+  protected
+
+  def initialize_service
+    @service_object = CrowbarService.new logger
+  end
+
+  def permitted_params
+    params.require(:crowbar)
   end
 end
-
