@@ -1,20 +1,19 @@
+# -*- encoding : utf-8 -*-
+#
 # Copyright 2011-2013, Dell
-# Copyright 2013, SUSE LINUX Products GmbH
+# Copyright 2013-2014, SUSE LINUX Products GmbH
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#  http://www.apache.org/licenses/LICENSE-2.0
+#   http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-#
-# Author: Rob Hirschfeld
-# Author: SUSE LINUX Products GmbH
 #
 
 class ProposalObject < ChefObject
@@ -82,7 +81,7 @@ class ProposalObject < ChefObject
 
   def pretty_attributes
     @pretty_attributes ||= begin
-      Utils::ExtendedHash.new(
+      Hashie::Mash.new(
         raw_attributes.dup
       )
     end
@@ -104,7 +103,7 @@ class ProposalObject < ChefObject
 
   def pretty_deployment
     @pretty_deployment ||= begin
-      Utils::ExtendedHash.new(
+      Hashie::Mash.new(
         raw_data["deployment"][barclamp].dup
       )
     end
@@ -132,12 +131,22 @@ class ProposalObject < ChefObject
 
   def name
     match = @item.name.match(/crowbar_bc-(.*)-(.*)$/)
-    match[1] == 'template' ? match[1] : match[2]
+
+    if match
+      match[1] == 'template' ? match[1] : match[2]
+    else
+      @item.name
+    end
   end
 
   def barclamp
     match = @item.name.match(/crowbar_bc-(.*)-(.*)$/)
-    match[1] == 'template' ? match[2] : match[1]
+
+    if match
+      match[1] == 'template' ? match[2] : match[1]
+    else
+      @item.name
+    end
   end
 
   def prop
