@@ -1,8 +1,27 @@
-maintainer       "User Unknown"
-maintainer_email "Unknown@Sample.com"
-license          "Apache 2.0"
-description      "Installs/Configures ==*BC-MODEL=="
-long_description IO.read(File.join(File.dirname(__FILE__), 'README.me'))
-version          "0.0"
+# -*- encoding : utf-8 -*-
+#
+# Copyright 2011-2013, Dell
+# Copyright 2013-2014, SUSE LINUX Products GmbH
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+maintainer "Unknown User"
+maintainer_email "unknown@sample.com"
+license "Apache 2.0"
+description "Installs/Configures ==*BC-MODEL=="
+long_description IO.read(File.join(File.dirname(__FILE__), "README.md"))
+version "0.0"
 
 depends "nagios"
+
