@@ -1,10 +1,13 @@
-# Copyright 2011 Dell, Inc.
+# -*- encoding : utf-8 -*-
+#
+# Copyright 2011-2013, Dell
+# Copyright 2013-2014, SUSE LINUX Products GmbH
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#     http://www.apache.org/licenses/LICENSE-2.0
+#   http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -12,6 +15,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
+
 package "==BC-MODEL==" do
   options "--force-yes"
   action :install
@@ -23,3 +27,4 @@ template node[:==BC-MODEL==][:config_file] do
   group "root"
   mode 0644
 end
+
