@@ -1,11 +1,13 @@
+# -*- encoding : utf-8 -*-
+#
 # Copyright 2011-2013, Dell
-# Copyright 2013, SUSE LINUX Products GmbH
+# Copyright 2013-2014, SUSE LINUX Products GmbH
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#  http://www.apache.org/licenses/LICENSE-2.0
+#   http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,37 +15,36 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-# Author: Rob Hirschfeld
-# Author: SUSE LINUX Products GmbH
-#
 
-# In the development environment your application's code is reloaded on
-# every request. This slows down response time but is perfect for development
-# since you don't have to restart the webserver when you make code changes.
-config.cache_classes = false
+Crowbar::Application.configure do
+  config.reload_classes_only_on_change = false
 
-# Full error reports are enabled and caching is turned off
-config.action_controller.consider_all_requests_local = true
-config.action_controller.perform_caching = false
-config.action_view.cache_template_loading = false
-config.action_view.debug_rjs = true
+  config.cache_classes = false
+  config.eager_load = false
+  config.consider_all_requests_local = true
 
-# Log error messages when you accidentally call methods on nil.
-config.whiny_nils = true
+  config.log_level = :debug
 
-# Set a verbose log level to get the informations we need
-config.log_level = :debug
+  config.i18n.fallbacks = false
 
-# Don't care if the mailer can't send
-config.action_mailer.raise_delivery_errors = false
+  config.serve_static_assets = true
+  config.assets.debug = true
+  config.assets.js_compressor = :closure
+  config.assets.css_compressor = :sass
+  config.assets.compile = true
+  config.assets.digest = false
+  config.assets.version = "1.0"
 
-# Enable request forgery protection in development environment
-config.action_controller.allow_forgery_protection = true
+  config.action_dispatch.show_exceptions = true
+  config.action_controller.perform_caching = false
+  config.action_controller.allow_forgery_protection = true
+  config.action_mailer.raise_delivery_errors = false
+  config.active_support.deprecation = :log
+  config.active_record.migration_error = :page_load
 
-# Enable threaded mode
-#config.threadsafe!
+  config.logger = ActiveSupport::TaggedLogging.new(
+    Logger.new Rails.root.join("log", "development.log") # SyslogLogger.new
+  )
 
-CHEF_CLIENT_KEY = "/opt/dell/crowbar_framework/config/client.pem"
-CHEF_NODE_NAME ="crowbar"
-CHEF_SERVER_URL = "http://localhost:4000"
-CROWBAR_VERSION = "Development"
+  #config.logger.formatter = ::Logger::Formatter.new
+end
