@@ -1,11 +1,13 @@
+# -*- encoding : utf-8 -*-
+#
 # Copyright 2011-2013, Dell
-# Copyright 2013, SUSE LINUX Products GmbH
+# Copyright 2013-2014, SUSE LINUX Products GmbH
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#  http://www.apache.org/licenses/LICENSE-2.0
+#   http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,40 +15,35 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-# Author: Rob Hirschfeld
-# Author: SUSE LINUX Products GmbH
-#
 
-# The test environment is used exclusively to run your application's
-# test suite.  You never need to work with it otherwise.  Remember that
-# your test database is "scratch space" for the test suite and is wiped
-# and recreated between test runs.  Don't rely on the data there!
-config.cache_classes = true
+Crowbar::Application.configure do
+  config.cache_classes = true
+  config.eager_load = false
+  config.consider_all_requests_local = true
 
-# Full error reports are disabled and caching is turned on
-config.action_controller.consider_all_requests_local = true
-config.action_controller.perform_caching = false
-config.action_view.cache_template_loading = true
-config.action_view.debug_rjs = false
+  config.log_level = :info
 
-# Log error messages when you accidentally call methods on nil.
-config.whiny_nils = true
+  config.i18n.fallbacks = false
 
-# Set a verbose log level to get the informations we need
-config.log_level = :debug
+  config.serve_static_assets = true
+  config.assets.debug = true
+  config.assets.js_compressor = :closure
+  config.assets.css_compressor = :sass
+  config.assets.compile = true
+  config.assets.digest = false
+  config.assets.version = "1.0"
 
-# Tell Action Mailer not to deliver emails to the real world.
-# The :test delivery method accumulates sent emails in the
-# ActionMailer::Base.deliveries array.
-config.action_mailer.delivery_method = :test
+  config.action_dispatch.show_exceptions = false
+  config.action_controller.perform_caching = false
+  config.action_controller.allow_forgery_protection = false
+  config.action_mailer.raise_delivery_errors = false
+  config.active_support.deprecation = :stderr
+  config.active_record.migration_error = :page_load
+  config.action_mailer.delivery_method = :test
 
-# Disable request forgery protection in test environment
-config.action_controller.allow_forgery_protection = false
+  config.logger = ActiveSupport::TaggedLogging.new(
+    Logger.new Rails.root.join("log", "test.log") # SyslogLogger.new
+  )
 
-# Enable threaded mode
-#config.threadsafe!
-
-CHEF_CLIENT_KEY = "/opt/dell/crowbar_framework/config/client.pem"
-CHEF_NODE_NAME ="crowbar"
-CHEF_SERVER_URL = "http://localhost:4000"
-CROWBAR_VERSION = "Test"
+  #config.logger.formatter = ::Logger::Formatter.new
+end
