@@ -1,5 +1,29 @@
-require "chef"
+# -*- encoding : utf-8 -*-
+#
+# Copyright 2011-2013, Dell
+# Copyright 2013-2014, SUSE LINUX Products GmbH
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
 
-Chef::Config.node_name CHEF_NODE_NAME
-Chef::Config.client_key CHEF_CLIENT_KEY
-Chef::Config.chef_server_url CHEF_SERVER_URL
+begin
+  require "chef"
+
+  Chef::Config.tap do |config|
+    config.node_name ENV["CHEF_NODE_NAME"]
+    config.client_key ENV["CHEF_CLIENT_KEY"]
+    config.chef_server_url ENV["CHEF_SERVER_URL"]
+  end
+rescue LoadError => e
+  Rails.logger.warn "Failed to load chef"
+end
