@@ -1,4 +1,29 @@
-SimpleNavigation::Configuration.run do |navigation|  
+# -*- encoding : utf-8 -*-
+#
+# Copyright 2011-2013, Dell
+# Copyright 2013-2014, SUSE LINUX Products GmbH
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+SimpleNavigation::Configuration.run do |navigation|
+  navigation.renderer = SimpleNavigationRenderers::Bootstrap3
+  navigation.consider_item_names_as_safe = true
+
+  navigation.selected_class = "active"
+  navigation.active_leaf_class = "leaf"
+
   navigation.items do |primary|
+    primary.dom_class = "nav navbar-nav"
   end
 end
