@@ -1,4 +1,22 @@
-require 'singleton'
+# -*- encoding : utf-8 -*-
+#
+# Copyright 2011-2013, Dell
+# Copyright 2013-2014, SUSE LINUX Products GmbH
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+require "singleton"
 
 class BarclampCatalog
   include Singleton
@@ -10,7 +28,7 @@ class BarclampCatalog
   attr_reader :categories, :catalog
 
   def initialize
-    @catalog    = YAML.load_file(Rails.root.join('config', 'catalog.yml'))
+    @catalog = YAML.load_file(Rails.root.join("config", "catalog.yml"))
     @categories = build_categories
   end
 
@@ -19,23 +37,25 @@ class BarclampCatalog
   end
 
   def barclamps
-    catalog['barclamps']
+    catalog["barclamps"]
   end
 
   def chef_order(barclamp)
     return DEFAULT_CHEF_ORDER if barclamp.nil?
+
     order = (barclamps[barclamp]["order"] || DEFAULT_CHEF_ORDER) rescue DEFAULT_CHEF_ORDER
     (barclamps[barclamp]["chef_order"] || order) rescue order
   end
 
   def run_order(barclamp)
     return DEFAULT_RUN_ORDER if barclamp.nil?
+
     order = (barclamps[barclamp]["order"] || DEFAULT_RUN_ORDER) rescue DEFAULT_RUN_ORDER
     (barclamps[barclamp]["run_order"] || order) rescue order
   end
 
   def members(barclamp)
-    barclamps[barclamp].nil? ? [] : barclamps[barclamp]['members']
+    barclamps[barclamp].nil? ? [] : barclamps[barclamp]["members"]
   end
 
   def category(barclamp)
@@ -47,7 +67,7 @@ class BarclampCatalog
   end
 
   def display_name(barclamp)
-    display = barclamps[barclamp]['display']
+    display = barclamps[barclamp]["display"] rescue nil
 
     if display.nil? or display.empty?
       barclamp.titlecase
@@ -67,3 +87,4 @@ class BarclampCatalog
     end
   end
 end
+
