@@ -1,3 +1,21 @@
+# -*- encoding : utf-8 -*-
+#
+# Copyright 2011-2013, Dell
+# Copyright 2013-2014, SUSE LINUX Products GmbH
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
 module Dsl
   module Proposal
     class Attribute
@@ -31,7 +49,7 @@ module Dsl
           :type => "hidden",
           :name => "proposal_attributes",
           :value => attrs.to_json
-        )
+        ).html_safe
       end
 
       def select_field(attribute, options = {})
@@ -56,7 +74,7 @@ module Dsl
                 "data-change" => changer("string")
               }).merge(options)
             )
-          ].join("\n")
+          ].join("\n").html_safe
         end
       end
 
@@ -86,7 +104,7 @@ module Dsl
                 "data-change" => changer("boolean")
               }).merge(options)
             )
-          ].join("\n")
+          ].join("\n").html_safe
         end
       end
 
@@ -149,7 +167,7 @@ module Dsl
               ),
               :class => "pull-right"
             )
-          ].join("\n")
+          ].join("\n").html_safe
         )
       end
 
@@ -202,7 +220,7 @@ module Dsl
           [
             labeling,
             send(field, *parameters)
-          ].join("\n")
+          ].join("\n").html_safe
         end
       end
 
@@ -211,7 +229,7 @@ module Dsl
           :div,
           yield,
           { :class => "form-group" }.merge(options || {})
-        )
+        ).html_safe
       end
 
       def defaults
@@ -228,7 +246,7 @@ module Dsl
           :label,
           t(translation_key.join(".")),
           :for => attribute
-        )
+        ).html_safe
       end
 
       def changer(type_cast)
