@@ -1,3 +1,21 @@
+# -*- encoding : utf-8 -*-
+#
+# Copyright 2011-2013, Dell
+# Copyright 2013-2014, SUSE LINUX Products GmbH
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
 module Dsl
   module Proposal
     class Deployment
@@ -18,11 +36,11 @@ module Dsl
       def deployments_field
         tag(
           :input,
-          :id => "proposal_attributes",
+          :id => "proposal_deployment",
           :type => "hidden",
-          :name => "proposal_attributes",
+          :name => "proposal_deployment",
           :value => attrs.to_json
-        )
+        ).html_safe
       end
 
       def header(dep_raw = true, attr_raw = false)
@@ -38,7 +56,7 @@ module Dsl
               ),
               :class => "pull-right"
             )
-          ].join("\n")
+          ].join("\n").html_safe
         )
       end
 
