@@ -1,3 +1,5 @@
+# -*- encoding : utf-8 -*-
+#
 # Copyright 2009, Opscode, Inc.
 # Copyright 2008, Steve Midgley
 #
@@ -5,16 +7,13 @@
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#  http://www.apache.org/licenses/LICENSE-2.0
+#   http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-#
-# Author: Adam Jacob
-# Author: Steve Midgley
 #
 
 # Copied from https://github.com/opscode/chef/blob/master/lib/chef/mixin/deep_merge.rb
