@@ -1,7 +1,25 @@
-require 'spec_helper'
+# -*- encoding : utf-8 -*-
+#
+# Copyright 2011-2013, Dell
+# Copyright 2013-2014, SUSE LINUX Products GmbH
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+require "spec_helper"
 
 describe CrowbarController do
-  integrate_views
+  render_views
 
   before do
     CrowbarService.any_instance.stubs(:apply_role).returns([200, "OK"])
@@ -113,7 +131,7 @@ describe CrowbarController do
       it "redirects to propsal path on failure" do
         CrowbarService.any_instance.stubs(:show_active).returns([500, "Error"])
         get :show, :id => "default"
-        response.should redirect_to(proposal_barclamp_path(:controller => "crowbar", :id => "default"))
+        response.should redirect_to(proposal_show_path(:controller => "crowbar", :id => "default"))
       end
     end
   end
@@ -252,3 +270,4 @@ describe CrowbarController do
     end
   end
 end
+
