@@ -1,6 +1,23 @@
+# -*- encoding : utf-8 -*-
+#
+# Copyright 2011-2013, Dell
+# Copyright 2013-2014, SUSE LINUX Products GmbH
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
 silence_warnings do
-  # This will make chef silent about missing key to sign requests
-  CHEF_CLIENT_KEY = nil
+  ENV["CHEF_CLIENT_KEY"] = nil
 end
 
-Chef::Config.client_key CHEF_CLIENT_KEY
+Chef::Config.client_key ENV["CHEF_CLIENT_KEY"]
